requirements.txt
langchain==0.3.26
pydantic-settings==2.10.1
starlette==0.47.2
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
langchain-google-genai==2.1.8
google-generativeai==0.8.5
notebook==7.4.4
PyPDF2==3.0.1
langchain-chroma==0.2.5